pipeline/requirements.txt
pymongo>=4.10.0
motor>=3.6.0
pydantic-settings>=2.6.0
python-dotenv>=1.0.0
pyyaml>=6.0.0
pandas>=2.2.0
httpx>=0.28.0
tqdm>=4.67.0
openpyxl>=3.1.0
pycountry>=24.6.0
beautifulsoup4>=4.13.0
lxml>=5.0.0
requests>=2.32.0
pytest>=8.3.0
mongomock>=4.1.0
pytest-asyncio>=0.24.0